Some more cleanups and removing the per capita default.
