+++ server/config/chart_config/crime_new.json
@@ -1,10 +1,10 @@
 [{
-    "category": "Economics",
-    "titleId": "CHART_TITLE-Gdp",
-    "title": "GDP",
-    "description": "GDP",
+    "category": "Crime",
+    "titleId": "CHART_TITLE-Total_crime",
+    "title": "Total crime",
+    "description": "Total number of criminal incidents",
     "variables": [
-      "Amount_EconomicActivity_GrossDomesticProduction_Nominal"
+      "Count_CriminalActivities_CombinedCrime"
     ],
     "denominator": [],
     "nonDividable": true,
@@ -43,7 +43,7 @@
               },
               {
                 "type": "RANKING",
-                "maxPlaces": 5
+                "maxPlaces": 10
               }
             ]
           },
@@ -53,7 +53,7 @@
             "charts": [
               {
                 "type": "BAR",
-                "maxPlaces": 5
+                "maxPlaces": 10
               }
             ]
           },
