+++ static/js/shared/util.ts
@@ -29,7 +29,10 @@ export const placeExplorerCategories = [
   "housing",
   "environment",
   "energy",
+  "health_new",
   "energy_new",
+  "crime_new",
+  "demographics_new",
   "economics_new",
 ];
